+++ CHANGELOG.md
@@ -9,7 +9,7 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
   * Added Smart Build Divider Tool
   * Added Smart Build Front Panel Tool
   * Added Smart Build Back Panel Tool
-  * Added Smart Build Module Tool **(DEV team only)**
+  * Added Smart Build Module Tool
   * Added Smart Stretch Tool
   * Added Smart Paneling Tool
   * Added Smart Boolean Tools
@@ -30,6 +30,9 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2026-09-25
+ * Improved Smart Build Tools
+ * Updated PackingSolver dependency
 ### 2026-09-24
  * Improved Smart Build Module Tool **(DEV team only)**
 ### 2026-09-23